requirements.txt
djangorestframework
markdown
django-filter

# git bash 
# source .venv/Scripts/activate